// apps/bridge/codegen/model/onrampQuoteProvider.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * API
 * API docs
 * OpenAPI spec version: 1.0
 */

export type OnrampQuoteProvider = typeof OnrampQuoteProvider[keyof typeof OnrampQuoteProvider];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const OnrampQuoteProvider = {
  ramp: 'ramp',
  Transak: 'Transak',
  Halliday: 'Halliday',
  Moonpay: 'Moonpay',
} as const;